pyproject.toml
[project]
name = "arsPRJ"
version = "0.1.0"
description = "Arsenal data ETL"
requires-python = ">=3.9"
dependencies = [
  "requests",
  "pandas",
  "psycopg2-binary",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "batching/code"}

[tool.setuptools.packages.find]
where = ["batching/code"]
